Canvas as floats

+++ rust/src/color.rs
@@ -5,16 +5,15 @@ use rando::{rand, rand_adjust, rand_color_adjust};
 use std::hash::{Hash, Hasher};
 
 #[inline]
-fn color_add(c:u8, c2: u8, opacity: f32) -> u8 {
-	return (c as f32 * (1. - opacity) +
-           (c2 as f32 * opacity)) as u8;
+fn color_add(c:f32, c2: f32, opacity: f32) -> f32 {
+	return c * (1. - opacity) + (c2 * opacity);
 }
 
 #[derive(PartialEq, Debug, Clone, Serialize, Deserialize)]
 pub struct Color {
-    pub r: u8,
-    pub g: u8,
-    pub b: u8,
+    pub r: f32, // 0 - 255 (u8)
+    pub g: f32,
+    pub b: f32,
     pub opacity: f32
 }
 
@@ -63,14 +62,14 @@ impl Color {
     }
 
     #[inline]
-    pub fn add_to_vec(&self, vec: &mut Vec<u8>, i: usize){
+    pub fn add_to_vec(&self, vec: &mut Vec<f32>, i: usize){
         vec[i]     = color_add(vec[i],      self.r, self.opacity);
         vec[i + 1] = color_add(vec[i + 1],  self.g, self.opacity);
         vec[i + 2] = color_add(vec[i + 2],  self.b, self.opacity);
     }
 
     pub fn black() -> Color {
-        Color {r:0,g:0,b:0,opacity:1.}
+        Color {r:0.,g:0.,b:0.,opacity:1.}
     }
 }
 
@@ -78,7 +77,7 @@ impl fmt::Display for Color {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
     return write!(f,
             "rgba({},{},{},{:.4})",
-            self.r, self.g, self.b, self.opacity)
+            self.r as u8, self.g as u8, self.b as u8, self.opacity)
     }
 }
 
@@ -100,9 +99,9 @@ impl Mul<f32> for Color {
 
     fn mul(self, _rhs: f32) -> Color {
         Color {
-            r: (self.r as f32 * _rhs) as u8,
-            g: (self.g as f32 * _rhs) as u8,
-            b: (self.b as f32 * _rhs) as u8,
+            r: self.r * _rhs,
+            g: self.g * _rhs,
+            b: self.b * _rhs,
             opacity: self.opacity * _rhs
         }
     }
@@ -110,9 +109,9 @@ impl Mul<f32> for Color {
 
 impl Hash for Color {
     fn hash<H: Hasher>(&self, state: &mut H) {
-        self.r.hash(state);
-        self.g.hash(state);
-        self.b.hash(state);
+        (self.r as u8).hash(state);
+        (self.g as u8).hash(state);
+        (self.b as u8).hash(state);
         ((self.opacity * 1000.) as i32).hash(state);
     }
 }
@@ -123,7 +122,7 @@ mod tests {
     
     #[test]
     fn test_color_add() { 
-        let c = color_add(255, 255, 1.);
+        let c = color_add(255., 255., 1.);
         assert_eq!(c, 255);
     }
 }
